* libopenvas/store.c (safe_copy): Cast return value strlen to match format string.  Silences a compiler warning
git-svn-id: svn+ssh://scm.wald.intevation.org/openvas/trunk/openvas-libraries@651 423fd1db-d629-0410-8442-d21db03e70f4

+++ libopenvas/store.c
@@ -113,7 +113,7 @@ static int safe_copy(char * str, char * dst, int sz, char * path, char * item)
   
  if(strlen(str) >= sz)
  {
-  fprintf(stderr, "openvas-libraries/libopenvas/store.c: %s has a too long %s (%ld)\n", path, item, strlen(str));
+  fprintf(stderr, "openvas-libraries/libopenvas/store.c: %s has a too long %s (%ld)\n", path, item, (long)strlen(str));
   return -1;
  }
  strcpy(dst, str);
